+++ CHANGELOG.rst
@@ -33,6 +33,8 @@ New features and enhancements
     uses one of such arguments with a ``cls`` that doesn't support them, Cloup
     augments the resulting ``TypeError`` with extra information.
 
+- In dark and light themes, the epilog is now left unstyled by default. :issue:`62`
+
 Bug fixes
 ---------
 - ``SectionMixin.add_section`` called ``super().add_command`` rather than